Last Fix

+++ BAR/Database/Entities/Balance.cs
@@ -3,7 +3,7 @@ namespace BAR.Database.Entities;
 public class Balance
 {
     public string AccountId { get; set; } = string.Empty;
-    public int Amount { get; set; } = 100;
+    public long Amount { get; set; } = 100;
 
     public Account? Account { get; set; }
 }
